Overhaul of packages and exported surface area (#40)
This removes many items from the exported API surface area.
Things that were previously configurable/extensible are now
hard-coded, such as the definition of "warm", how health
check results are interpreted to compute "usable" connections,
and how connections are managed/reconciled in the face of
resolver updates.

The "balancer" itself is also no longer exposed. Instead, only
the picker and resolver can be customized. Subsetting can be
implemented via decorating a resolver. Internally, the
abstractions of "balancer" and "connection manager" remain,
but they are unexported types used inside of the transport.
These abstractions remain primarily to increase testability and
make sure we still have adequate testing of the various
components of the transport.