Fix heap update operation for least-loaded picker (#61)
The `update` method of the heap, which reconciles the set of connections
with a new set of connections provided from a resolver, was sometimes
calling `heap.Remove`, which not only would change the order of elements
making the iteration invalid, but also shorten the slice, making the iteration
eventually panic with an index-out-of-bounds failure.

This changes the logic to not use `heap.Remove` during the iteration.
Instead, it mutates the slice, which potentially breaks heap invariants. So
instead of `heap.Push` for new items, it simply appends them, and then it
does `heap.Init` to re-heapify the slice at the end.

This also adds a new test to verify that the heap is working under a variety
of scenarios, including updates that may need to add and remove connections.