Eliminate test flakes in recently-added heap tests (#65)
The update operation includes iteration over a map, so the
order of items picked from the heap is not entirely deterministic.
This fixes the tests to be lax on the precise order of elements
when they their load is equal/tied. The thing we care about is
that items aren't picked when their load is higher than others.