Implement address family affinity and prefer IPv4 by default (#66)
Implements address family policy. When the policy is set to `PreferIPv4`
or `PreferIPv6`, any DNS result that has both IPv4 and IPv6 records will
be filtered to only the preferred records; in cases where only IPv4 or
IPv6 addresses are present, this has no impact. The default is set to
`PreferIPv4` to help lower the likelihood of problems due to the fact
that `httplb` lacks some kind of implementation of a "happy eyeballs"
algorithm to perform IPv6 fallback.

It winds up being a little tricky to test this, but it's probably
worthwhile; right now we can only test the DNS resolver in a fairly
limited way (by relying on the fact that resolving an IP address is a
no-op.) With this approach, we can test the DNS resolver more-or-less
end-to-end.