Update usage of random number generators (#77)
Go 1.22 included a new "math/rand/v2" package (see the [blog
post](https://go.dev/blog/randv2) for more details). The most exciting
thing about this is a new global RNG that uses per-thread seeds and is
lock-free. And it turns out that this approach is indeed _much_ better
than using a mutex with a `*rand.Rand`.

But for random numbers that **don't** need to be thread-safe, the Go 1
generator in "math/rand" is still the fastest.

So this PR removes `internal.NewLockedRand()` and updates all callers to
instead use the global RNG in "math/rand/v2".