Add go1.24-specific implementation of "h2c" scheme (#76)
As of Go 1.24, we no longer need to import "golang.org/x/net/http2" in
order to support the "h2c" scheme (to force HTTP/2 over plain-text).

This adds a Go-1.24-specific implementation that uses the latest H2C
features in "net/http". Build tags are used so that we continue using
the old implementation for projects still using Go 1.23.