Push enforcement of min refresh interval to resolver (#84)
This was previously handled in the balancer, but a rate limit for
resolving is really more of a resolver responsibility.

This also fixes an issue in the code: previously, if an attempt to
"resolve now" would exceed the rate limit, the "resolve now" intent
would be lost (instead of applying back-pressure). So the resolver
wouldn't actually be refreshed until the next scheduled TTL-based
refresh. Now, if the rate limit is exceeded, the resolver will wait the
minimum refresh lapses.

While making this change and the above fix, I also expose the minimum
refresh interval for callers to configure, to resolve #83. This also
allows a resolver to put a lower bound on TTLs, which could
independently be useful to avoid spamming the prober due to
misconfigured/pathological TTLs in the results.

This also updates to a newer version of the fake clock package, which
improves a few things and allowed some cleanup in our internal fake
clock wrapper.